Move test-workspace out of the context of the e2e-tests project.
The e2e tests run with mocha and Stryker is being tested using a Jest runner, the test-workspace tests also use Jest. Having Jest and Mocha installed in the same project caused conflict, simplest way to ensure the integration tests still run now that the extension will run in the context of the nearest package.json is to move them to a different project.